update vmin and vmax only if it narrows down possible versions
This a) allows one to just comment out "break" to execute all tests without falsifying possible versions and
b) does preparations for future bug tests to not intervene with non-configurable features.

If we detect a non-present, non-configurable feature, we set vmax. With this patch, we prevent to override vmax
with a greater vmax of a positive bug test.