execute feature test for MSR IA32_ARCH_CAPABILITIES on Intel only
Support for this MSR via CPUID was added in 4.15 for Intel only [1].
4.17 added CPUID support for all platforms [2], but we already have
a reliable test for that version.

Note that [2] only applies to CPUID support. Proper MSR emulation
for AMD was first added in 5.1 [3].
A bug test could probably be created from that, but would work only
on AMD platforms.

[1]: https://git.kernel.org/pub/scm/virt/kvm/kvm.git/commit/?id=28c1c9fabf48d6ad596273a11c46e0d0da3e14cd
[2]: https://git.kernel.org/pub/scm/virt/kvm/kvm.git/commit/?id=1eaafe91a0df4157521b6417b3dd8430bf5f52f0
[3]: https://github.com/torvalds/linux/commit/0cf9135b773bf32fba9dd8e6699c1b331ee4b749