fix version of bug test MSR_K8_TSEG_ADDR
The introducing version of the bug fix is 4.2.4 [1].
However, Hyperprobe does not support patch versions yet, in terms of
semantic versioning.

Defining 4.2 as the fixed version leads to inaccurate results, because
if the bug is present, max. 4.1 is assumed, which is wrong.

Defining 4.3 leads to max. 4.2 if bug is present, which is also not
100% correct, but tests in Hyperprobe are intended to only have an
accuracy of minor versions anyway.

Should this bug test be used for assuming a min. version in the future,
it would assume min. 4.3. This is not as accurate as possible, but at
least also not wrong.

Support for patch versions would probably require to rework the whole
versioning system to support semantic versioning, but minor versions
are accurate enough for now.

[1]: https://elixir.bootlin.com/linux/v4.2.4/A/ident/MSR_K8_TSEG_MASK